Document song-level decoding; add a START HERE orientation section
The headline finding: per-window accuracy, which most of this session
optimised, is the wrong metric. Phase advances deterministically, so one
number determines a whole song's labelling and every beat casts a vote for
one of eight hypotheses. Errors disperse across the seven wrong hypotheses
while correct answers concentrate, so a model that is 0.686 accurate per
window decodes 12 of 16 held-out songs exactly right. Mean song-level
accuracy 0.894 against 0.686 per-window.

It fails on exactly the songs where errors do NOT disperse. Amor y Control
puts 556 votes on the hypothesis four away from the truth against 399 on the
truth, so aggregation amplifies the 1-5 flip and the song decodes to 0.000
from a per-window 0.408 -- the behaviour section 4 predicted analytically
long before there was a model to test it on.

Errors are also bursty rather than salt-and-pepper: easy songs run 29
consecutive correct windows broken by single slips, while Amor y Control's
wrong runs are longer than its right ones. Phase is established in stretches
and the model carries nothing between windows, which is why the online filter
(0.800) sits between raw (0.686) and batch (0.894).

Adds a START HERE section with current state, open questions, the traps
already hit, and how to run each script, since the context and architecture
sweeps in sections 6-9 were all measured on the proxy metric and have not
been re-run at song level. Also renumbers the sections, which had three
competing 9s from successive insertions, and fixes the cross-references.

+++ stage_b_eval_online.py
@@ -39,7 +39,11 @@ def posteriors(model, song, W, bs=256):
     with torch.no_grad():
         for start in range(W - 1, nb, bs):
             idxs = range(start, min(start + bs, nb))
-            X = np.stack([f[(b - W + 1) * FPB:(b + 1) * FPB] for b in idxs])
+            # load_all() keeps features float16 to save memory; batches()
+            # casts on assignment but np.stack preserves the dtype, so cast
+            # explicitly or conv2d rejects Half input against float weights.
+            X = np.stack([f[(b - W + 1) * FPB:(b + 1) * FPB]
+                          for b in idxs]).astype(np.float32)
             out.append(F.softmax(
                 model(torch.from_numpy(X).unsqueeze(1).to(DEV)), 1).cpu().numpy())
     return np.concatenate(out)
